requirements.txt
requests>=2.25.1
python-dotenv>=0.19.2
playwright>=1.30.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.1.0
streamlit>=1.28.0
pandas>=1.3.0
plotly>=5.0.0
# LangGraph dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5